[Android-Crypto] AddDigitalSign data layer

+++ android-app/app/src/main/java/com/bintangfajarianto/gmayl/data/repository/crypto/CryptoNetworkRepository.kt
@@ -1,8 +1,10 @@
 package com.bintangfajarianto.gmayl.data.repository.crypto
 
+import java.math.BigInteger
+
 interface CryptoNetworkRepository {
     suspend fun decryptMail(hexBody: String, symmetricKey: String): String
     suspend fun generateKeyPair(): Pair<String, String>
-    suspend fun sign(privateKey: String, message: String): Pair<Int, Int>
-    suspend fun verify(publicKey: String, message: String, r: Int, s: Int): Boolean
+    suspend fun sign(privateKey: String, message: String): Pair<BigInteger, BigInteger>
+    suspend fun verify(publicKey: String, message: String, r: BigInteger, s: BigInteger): Boolean
 }
